warn is feature complete; that was a bitch

+++ commands/mod/warn.js
@@ -1,8 +1,9 @@
 const commando = require('discord.js-commando');
+const Discord = require('discord.js');
 
-class WarnCommand extends commando.Command {
-  constructor(client) {
-    super(client, {
+module.exports = class WarnCommand extends commando.Command {
+  constructor(stevebot) {
+    super(stevebot, {
       name: 'warn',
       group: 'mod',
       memberName: 'warn',
@@ -10,11 +11,12 @@ class WarnCommand extends commando.Command {
     });
   }
 
-  async run(message, args) {
+  async run(message, stevebot, args) {
     var messageAuthor = message.guild.member(message.author);
     var warnedUser = message.mentions.users.first();
     var numOfMentions = message.mentions.users.size;
     var reason = message.content.split('\"');
+    var modlog = message.guild.channels.find('name', 'modlog');
     if (!messageAuthor.hasPermission("MANAGE_MESSAGES")) { //check if author of command message is mod or admin
       message.delete();
       return message.reply('you do not have this permission.');
@@ -24,14 +26,21 @@ class WarnCommand extends commando.Command {
     } else if (reason.length < 2) { //check that a reason has been given
       message.delete();
       return message.reply('you must supply a reason for the warning.');
-    } else if (reason.length > 2) { //check that only one set of quotes was used
+    } else if (reason.length > 2 && reason[2].length > 1) { //check that only one set of quotes was used
       message.delete();
       return message.reply('please only use one set of quotes for the reason.');
     }
       reason = reason[1];
       message.channel.sendMessage(warnedUser + ', please turn the bus around. ' + reason + ' :bus:');
       message.delete();
+      //modlog
+      const embed = new Discord.RichEmbed()
+      .setColor(0x00AE86)
+      .setTimestamp()
+      .addField('Action:', 'Warning')
+      .addField('User:', `${warnedUser.username}#${warnedUser.discriminator}`)
+      .addField('Modmin:', `${message.author.username}#${message.author.discriminator}`)
+      .addField('Reason:', reason);
+      modlog.sendEmbed(embed);
   }
 }
-
-module.exports = WarnCommand;
